Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-printf("Size of a char: %d byte(s)", (unsigned long)sizeof(char));
-printf("Size of an int: %d byte(S)", (unsigned long)sizeof(int));
-printf("Size of a long int: %d byte(S)", (unsigned long)sizeof(long int));
-printf("Size of a long long int: %d byte(s)", (unsigned long)sizeof(long long int));
-printf("Size of a float: %d byte(s)", (unsigned long)sizeof(float));
+printf("Size of a char: %d byte(s)", sizeof(char));
+printf("Size of an int: %d byte(S)", sizeof(int));
+printf("Size of a long int: %d byte(S)", sizeof(long int));
+printf("Size of a long long int: %d byte(s)", sizeof(long long int));
+printf("Size of a float: %d byte(s)", sizeof(float));
 return (0);
 }
